fixed mapping somewhat now easier to identify tests etc.
